CDCEvent and DeltaEvent separated to classes
