+++ compatibility_spark-3/src/test/scala/za/co/absa/hyperdrive/compatibility/impl/writer/delta/CDCEvent.scala
@@ -16,7 +16,6 @@
 package za.co.absa.hyperdrive.compatibility.impl.writer.delta
 
 import java.sql.Timestamp
-import scala.util.Try
 
 case class CDCEvent(id: String, value: String, timestamp: Timestamp, eventType: String)
 
@@ -27,25 +26,5 @@ object CDCEvent {
       line <- lines
       values = line.split(",").map(_.trim)
     } yield CDCEvent(values(0), values(1), Timestamp.valueOf(values(2)), values(3))
-  }
-}
-
-case class DeltaEvent(id: String, value: String, timestamp: Timestamp, eventType: String, _start_date: Timestamp, _end_date: Timestamp, _is_current: Boolean)
-
-object DeltaEvent {
-  def loadFromFile(path: String): Seq[DeltaEvent] = {
-    val lines = FileUtils.readFileLines(getClass.getResource(path).getPath)
-    for {
-      line <- lines
-      values = line.split(",").map(_.trim)
-    } yield DeltaEvent(
-      values(0),
-      values(1),
-      Timestamp.valueOf(values(2)),
-      values(3),
-      Timestamp.valueOf(values(4)),
-      Try(Timestamp.valueOf(values(5))).getOrElse(null),
-      values(6).toBoolean
-    )
   }
 }
